refactor(daemon): split append ingest helper
Problem: the live batch processor was already over its file budget, and adding bounded append flushing made the central daemon file harder to keep below the enforced ceiling. The generated-surface aggregate check also reported stale docs under python -m devtools even when individual checks passed.\n\nWhat changed: move append-only ingest persistence into polylogue.sources.live.append_ingest, keep the LiveBatchProcessor as the orchestration surface, refresh topology/agent generated files, and make render-all check mode sequential so verification is deterministic.\n\nVerification: pytest tests/unit/core/test_memory.py tests/unit/sources/test_live_batch_support.py tests/unit/sources/test_live_watcher.py -q --tb=short passed with 68 tests; pytest tests/unit/devtools/test_render_all.py -q --tb=short passed with 4 tests; python -m devtools verify --quick passed in 39.92s.
