+++ docs/topology-status.md
@@ -25,14 +25,14 @@ Generated by `devtools render-topology-status`. Reads `docs/plans/topology-targe
 | archive-semantic | — | 13 | 13 | 0 | 0 |
 | archive-session | — | 6 | 6 | 0 | 0 |
 | archive-viewport | — | 5 | 5 | 0 | 0 |
-| core-primitive | — | 15 | 15 | 0 | 0 |
+| core-primitive | — | 16 | 16 | 0 | 0 |
 
 ### Summary
 
-- **Stable** (no move scoped): 691
+- **Stable** (no move scoped): 692
 - **Kernel** (polylogue/ root): 10
 - **Primitives** (storage-root): 13
 - **TBD** (cell needs explicit assignment): 0
-- **Total declared**: 820
-- **Realized polylogue/**/*.py**: 820 files declared
+- **Total declared**: 822
+- **Realized polylogue/**/*.py**: 822 files declared
 <!-- END GENERATED: topology-status -->
