+++ tests/unit/sources/test_live_watcher.py
@@ -29,7 +29,6 @@ from polylogue.sources.live.batch import (
     _FullIngestResult,
     last_complete_newline_from_tail,
 )
-from polylogue.sources.live.batch_support import _AppendPlan, _AppendResult
 from polylogue.sources.live.cursor import CursorRecord, CursorStore
 from polylogue.storage.runtime import RawConversationRecord
 from tests.infra.frozen_clock import FrozenClock
@@ -854,61 +853,6 @@ def test_full_parse_progress_groups_bounds_small_files_by_bytes(
 
     assert sum(byte_size for _ in groups[0]) <= _SMALL_FULL_PARSE_PROGRESS_MAX_BYTES
     assert groups == [paths[:2], paths[2:4], paths[4:]]
-
-
-@pytest.mark.asyncio
-async def test_live_append_plans_flush_in_bounded_groups(
-    tmp_path: Path,
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    root = tmp_path / "sessions"
-    root.mkdir()
-    paths = [root / f"{index}.jsonl" for index in range(5)]
-    for path in paths:
-        path.write_text('{"type":"session_meta","payload":{"id":"bounded"}}\n', encoding="utf-8")
-    cursor = CursorStore(tmp_path / "live.sqlite")
-    polylogue = SimpleNamespace(archive_root=tmp_path, backend=SimpleNamespace(db_path=cursor._db_path))
-    processor = LiveBatchProcessor(
-        cast(Any, polylogue),
-        (WatchSource(name="codex", root=root),),
-        cursor=cursor,
-        parser_fingerprint="test-parser",
-    )
-    groups: list[list[Path]] = []
-
-    def fake_append_plan(path: Path, *, cursor: CursorRecord | None = None) -> _AppendPlan:
-        del cursor
-        return _AppendPlan(
-            path=path,
-            source_name="codex",
-            start_offset=0,
-            last_complete_newline=10,
-            stat_size=10,
-            st_dev=1,
-            st_ino=1,
-            mtime_ns=1,
-            payload=b"payload\n",
-            payload_hash="tail",
-            cursor_fingerprint="base",
-            bytes_read=10,
-        )
-
-    def fake_ingest_append_plans(plans: list[_AppendPlan]) -> _AppendResult:
-        groups.append([plan.path for plan in plans])
-        return _AppendResult(succeeded=list(plans), failed=[], worker_count=1)
-
-    monkeypatch.setattr(processor, "_append_plan", fake_append_plan)
-    monkeypatch.setattr(processor, "_ingest_append_plans", fake_ingest_append_plans)
-    monkeypatch.setattr(processor, "_converge_paths", lambda paths: (paths, 0.0, {}, []))
-    monkeypatch.setattr(processor, "_record_append_cursor", lambda plan: True)
-    monkeypatch.setattr(processor, "_record_convergence_outcome", lambda path, debts: None)
-    monkeypatch.setattr("polylogue.sources.live.batch._append_plan_group_ready", lambda plans: len(plans) >= 2)
-
-    metrics = await processor.ingest_files(paths, emit_event=False)
-
-    assert groups == [paths[:2], paths[2:4], paths[4:]]
-    assert metrics.append_file_count == 5
-    assert metrics.full_file_count == 0
 
 
 @pytest.mark.asyncio
